Update README.md
Your assignment must be typed to produce a PDF document a3.pdf, and a plain text document reservation.ddl (hand-written submissions are not acceptable). You may work on the assignment in groups of 1 or 2, and

submit a single assignment for the entire group on MarkUs. You must establish your group well before the due date byReservationsubmiting an incomplete, or even empty, submission (of course, we only grade the last submission before the due date/time).

If you do work in group of 2, be sure that both partners are deeply involved in all parts of your assignment. This assignment is short enough to be completed by a single person, and we expect you to be able to submit it even if your partnership dissolves before the due date.